content-understanding-agent/requirements.txt
azure-ai-projects
azure-ai-agents
azure-identity
azure-ai-inference
azure-storage-blob
python-dotenv
requests
